têm muito que fazer

+++ TP_PD_cliente_m1/src/pt/isec/pd/trabalhoPratico/model/classesComunication/Mgs_RegistarEditar_Conta.java
@@ -2,9 +2,9 @@ package pt.isec.pd.trabalhoPratico.model.classesComunication;
 
 public class Mgs_RegistarEditar_Conta extends Geral{
    private String nome, email, password;
-   private long num_estudante;// este aqui criei como long mas talvez usar outro tipo de dados seja preferível
+   private int num_estudante;// este aqui criei como long mas talvez usar outro tipo de dados seja preferível
 
-    public Mgs_RegistarEditar_Conta(String nome, String email, String password, long num_estudante, Message_types tipo) {
+    public Mgs_RegistarEditar_Conta(String nome, String email, String password, int num_estudante, Message_types tipo) {
         super(tipo);
         this.nome = nome;
         this.email = email;
@@ -24,7 +24,7 @@ public class Mgs_RegistarEditar_Conta extends Geral{
         return password;
     }
 
-    public long getNum_estudante() {
+    public int getNum_estudante() {
         return num_estudante;
     }
 }
